requirements.txt
# Core dependencies
anthropic>=0.39.0
pydantic>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
streamlit>=1.39.0
streamlit-js-eval>=0.1.7
pandas>=2.0.0
openpyxl>=3.1.0

# Database
psycopg2-binary>=2.9.0
bcrypt>=4.0.0

# Development dependencies
pytest>=8.0.0
pytest-cov>=4.0.0
